tests/requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
pyyaml==6.0.1
